Remove unnecessary edits to the makefile.
Turns out that we can leave the rustup installation in (as long as we
add the rustup dependency to shell.nix). My reason for doing this is
that I want to stray from the source material as little as possible,
so that you should be able to run everything regardless of whether you
use nix or not.